requirements.txt
certifi==2020.12.5
cffi==1.14.5
cryptography==3.4.7
Django==2.2.3
django-ranged-response==0.2.0
django-simple-captcha==0.5.11
Pillow==8.1.2
pycparser==2.20
PyMySQL==1.0.2
pytz==2021.1
six==1.15.0
sqlparse==0.4.1
wincertstore==0.2
